# docs/swagger.yaml
basePath: /api
definitions:
...
  dtos.CreateDistritoDTO:
    properties:
      municipio_id:
        type: string
      nome:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - municipio_id
    - nome
    type: object
...
  dtos.DistritoResponseDTO:
    properties:
      created_at:
        type: string
      id:
        type: string
      municipio_id:
        type: string
      nome:
        type: string
      updated_at:
        type: string
    type: object
  dtos.DistritoResultDTO:
    properties:
      created_at:
        type: string
      id:
        type: string
      municipio_id:
        type: string
      municipio_nome:
        type: string
      nome:
        type: string
      updated_at:
        type: string
    type: object
...
  dtos.UpdateDistritoDTO:
    properties:
      municipio_id:
        type: string
      nome:
        maxLength: 100
        minLength: 2
        type: string
    type: object
...
  utils.FieldError:
    properties:
      field:
        type: string
      label:
        type: string
      message:
        type: string
      tag:
        type: string
      value: {}
    type: object
  utils.Payload:
    properties:
      error:
        description: 'ex: validation_failed, conflict, internal'
        type: string
      fields:
        description: detalhes por campo
        items:
          $ref: '#/definitions/utils.FieldError'
        type: array
      message:
        description: resumo humanizado
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: suporte@empresa.com
    name: Equipa de Desenvolvimento
  description: API de Gestão de Recursos Humanos (Províncias, Municípios, Departamentos,
    etc.)
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://localhost:8080/terms
  title: HRMS API
  version: "1.0"
paths:
...
  /districts:
    get:
      description: Retorna uma lista com todos os distritos cadastrados
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.DistritoResponseDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Payload'
      summary: Lista todos os distritos
      tags:
      - Distritos
    post:
      consumes:
      - application/json
      description: Registra um novo distrito associado a um município
      parameters:
      - description: Dados do distrito
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateDistritoDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.DistritoResponseDTO'
        "400":
          description: Validation error with fields detail
          schema:
            $ref: '#/definitions/utils.Payload'
        "409":
          description: Conflito (chave única)
          schema:
            $ref: '#/definitions/utils.Payload'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/utils.Payload'
      summary: Cria um novo distrito
      tags:
      - Distritos
  /districts/{id}:
    delete:
      description: Exclui um distrito pelo ID
      parameters:
      - description: ID do distrito
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Payload'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Payload'
      summary: Remove um distrito
      tags:
      - Distritos
    get:
      description: Retorna os dados de um distrito específico
      parameters:
      - description: ID do distrito
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.DistritoResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Payload'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Payload'
      summary: Busca um distrito pelo ID
      tags:
      - Distritos
    put:
      consumes:
      - application/json
      description: Modifica o nome e/ou município de um distrito existente
      parameters:
      - description: ID do distrito
        in: path
        name: id
        required: true
        type: string
      - description: Dados atualizados
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateDistritoDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.DistritoResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Payload'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Payload'
      summary: Atualiza os dados de um distrito
      tags:
      - Distritos
  /districts/search:
    get:
      consumes:
      - application/json
      description: Busca distritos com texto, filtros e paginação
      parameters:
      - description: Texto de busca (nome do distrito)
        in: query
        name: search
        type: string
      - description: Filtro por nome do município
        in: query
        name: filter
        type: string
      - description: Número máximo de registros a retornar
        in: query
        name: limit
        type: integer
      - description: Número de registros a ignorar (para paginação)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.DistritoResultDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Payload'
      summary: Buscar distritos
      tags:
      - Distritos
...
schemes:
- http
swagger: "2.0"